An update on the playbook
Greeting prof.
I know during my call we ran this playbook to copy the ssh key and create an ansible user but we had an error. so I was just going through it today  and i find out what the problem was i did the corrections and it working good now. So  I'm  proposing a cross-check and if possible you can merge it. Thank you.

Name  Samuel Munoh of Class30